+++ jazkarta/shop/stripe.py
@@ -1,6 +1,7 @@
 # Make sure we don't import this module recursively via `import stripe`
 from __future__ import absolute_import
 
+import six
 import stripe
 
 from jazkarta.shop import config
@@ -13,29 +14,17 @@ def stripe_amount(dollars):
     return int(dollars * 100)
 
 
-def call_stripe(method, url, params):
+def call_stripe():
     if config.IN_PRODUCTION:
         api_key_setting = 'stripe_api_key_production'
     else:
         api_key_setting = 'stripe_api_key_dev'
-    api_key = get_setting(api_key_setting)
-    api = stripe.APIRequestor(api_key)
-    try:
-        return api.request(method, url, params)[0]
-    except stripe.StripeError as e:
-        errmsg = e.json_body['error']['message']
-        raise PaymentProcessingException(errmsg)
-    except Exception as e:
-        raise PaymentProcessingException(str(e))
-
-
-def process_refund(charge_id, refund_amount):
-    return call_stripe('post', '/v1/charges/{}/refunds'.format(charge_id), {
-        'amount': stripe_amount(refund_amount),
-    })
+    stripe.api_key = get_setting(api_key_setting)
 
 
 def process_interactive_payment(cart, card_token, contact_info):
+    call_stripe()
+
     # Calculate total
     amount = stripe_amount(cart.amount)
     assert amount > 0
@@ -49,10 +38,19 @@ def process_interactive_payment(cart, card_token, contact_info):
         metadata['ship_to'] = shipping
 
     # Process the charge using Stripe
-    return call_stripe('post', '/v1/charges', {
-        'amount': int(amount),
-        'currency': 'usd',
-        'card': card_token,
-        'description': cart.summary,
-        'metadata': metadata,
-    })
+    try:
+        result = stripe.Charge.create(
+            amount = int(amount),
+            currency = "usd",
+            source = card_token,
+            description = cart.summary,
+            metadata = metadata
+        );
+    except stripe.error.StripeError as e:
+        errmsg = e.json_body['error']['message']
+        raise PaymentProcessingException(errmsg)
+
+    except Exception as e:
+        raise PaymentProcessingException(six.text_type(e))
+
+    return result